backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain==0.3.9
langchain-google-genai==2.0.5
python-dotenv==1.0.1
httpx==0.27.2
pydantic>=2.10.3
